+++ datadog_checks_dev/datadog_checks/dev/tooling/templates/integration/check_only/{check_name}/datadog_checks/__init__.py
@@ -1,0 +1,2 @@
+{license_header}
+__path__ = __import__('pkgutil').extend_path(__path__, __name__)  # type: ignore