+++ datadog_checks_dev/datadog_checks/dev/tooling/templates/integration/check_only/{check_name}/tests/test_unit.py
@@ -1,0 +1,24 @@
+{license_header}
+
+from typing import Any, Callable, Dict  # noqa: F401
+
+from datadog_checks.base import AgentCheck  # noqa: F401
+from datadog_checks.base.stubs.aggregator import AggregatorStub  # noqa: F401
+from datadog_checks.dev.utils import get_metadata_metrics
+from datadog_checks.{check_name} import {check_class}
+
+
+def test_check(dd_run_check, aggregator, instance):
+    # type: (Callable[[AgentCheck, bool], None], AggregatorStub, Dict[str, Any]) -> None
+    check = {check_class}('{check_name}', {{}}, [instance])
+    dd_run_check(check)
+
+    aggregator.assert_all_metrics_covered()
+    aggregator.assert_metrics_using_metadata(get_metadata_metrics())
+
+
+def test_emits_critical_service_check_when_service_is_down(dd_run_check, aggregator, instance):
+    # type: (Callable[[AgentCheck, bool], None], AggregatorStub, Dict[str, Any]) -> None
+    check = {check_class}('{check_name}', {{}}, [instance])
+    dd_run_check(check)
+    aggregator.assert_service_check('{check_name}.can_connect', {check_class}.CRITICAL)